starting to reorganize in prep for more refactoring
